Delete comments on thread/spool delete

+++ threadit-api/Services/ThreadService.cs
@@ -8,10 +8,32 @@ namespace ThreaditAPI.Services
     {
         private readonly ThreadRepository threadRepository;
         private readonly CommentRepository commentRepository;
+        private readonly SpoolRepository spoolRepository;
         public ThreadService(PostgresDbContext context)
         {
             this.threadRepository = new ThreadRepository(context);
             this.commentRepository = new CommentRepository(context);
+            this.spoolRepository = new SpoolRepository(context);
+        }
+
+        private async Task<bool> IsDeleteAuthorized(string threadId, string userId) {
+            Models.Thread? thread = await this.threadRepository.GetThreadAsync(threadId);
+
+            if (thread == null) {
+                return false;
+            }
+
+            if (userId == thread.OwnerId) {
+                return true;
+            }
+
+            Spool? spool = await this.spoolRepository.GetSpoolAsync(thread.SpoolId);
+
+            if (spool == null) {
+                return false;
+            }
+
+            return spool.OwnerId == userId || spool.Moderators.Contains(userId);
         }
 
         public async Task<Models.Thread?> GetThreadAsync(string threadId)
@@ -161,8 +183,13 @@ namespace ThreaditAPI.Services
             return thread;
         }
 
-        public async Task DeleteThreadAsync(string threadId)
+        public async Task DeleteThreadAsync(string threadId, string userId)
         {
+            if (!await IsDeleteAuthorized(threadId, userId)) {
+                throw new Exception("User does not have permission to delete comment.");
+            }
+
+            await this.commentRepository.HardDeleteAllThreadCommentsAsync(threadId);
             await this.threadRepository.DeleteThreadAsync(threadId);
         }
     }
